Including support for loading 100e8 stars. This patch includes:
+ Removal of some deprecated code

+ Splitting of StarComponent into StarComponent and DeepStarComponent

+ Separation of named and unnamed stars

+ Support for both 16-byte per star and 32-byte per star data files

The changes are still experimental and might break some features or
introduce some bugs.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=852296

+++ kstars/skycomponents/typedef.h
@@ -38,7 +38,7 @@ class StarBlock;
 
 typedef quint32                        DrawID;
 typedef quint32                        UpdateID;
-typedef unsigned short                 Trixel;
+typedef unsigned int                   Trixel;
 typedef unsigned short                 BufNum;
 
 typedef QVector< SkyPoint*>            SkyList;
